fixing eval for SP
