+++ setup.py
@@ -106,7 +106,11 @@ def get_package_version():
 
 extras_require = {
     "flash-attn": ["flash-attn==2.7.4.post1"],
-    "ring-flash-attn": ["ring-flash-attn>=0.1.4", "yunchang==0.6.0"],
+    "ring-flash-attn": [
+        "flash-attn==2.7.4.post1",
+        "ring-flash-attn>=0.1.4",
+        "yunchang==0.6.0",
+    ],
     "deepspeed": [
         "deepspeed==0.16.4",
         "deepspeed-kernels",
